+++ CHANGELOG.md
@@ -71,3 +71,5 @@ Now plugin always must receive `params` object
 * -//-
 
 ### 1.0.4
+
+* Add optional field `botConfig`